# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pharmacy Management System API",
        "description": "Complete API documentation for Pharmacy Management System including admin, student, teacher, and payment modules",
        "contact": {
            "email": "support@pharmacy.edu"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.pharmacy.edu",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/admin/institutes": {
            "post": {
                "tags": [
                    "Admin - Master Data"
                ],
                "summary": "Get all institutes",
                "description": "Retrieve all institutes using fn_admin_getallinstitutes_v1",
                "operationId": "57862ffe96cdba77fcf5010b6f724685",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "admin_user_id"
                                ],
                                "properties": {
                                    "admin_user_id": {
                                        "type": "integer",
                                        "example": 668
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/admin/semesters": {
            "post": {
                "tags": [
                    "Admin - Master Data"
                ],
                "summary": "Get all semesters",
                "description": "Retrieves all semesters list filtered by semester type. Calls: fn_admin_getallsemesters(p_admin_user_id, p_semester_type)",
                "operationId": "3574bf5af47544e5307b18bd6929989f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "admin_user_id",
                                    "semester_type"
                                ],
                                "properties": {
                                    "admin_user_id": {
                                        "description": "Admin user ID",
                                        "type": "integer",
                                        "example": 668
                                    },
                                    "semester_type": {
                                        "description": "Semester type filter (1=Odd, 2=Even, etc.)",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Semesters list retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "version": {
                                            "type": "string",
                                            "example": "1.0"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully retrieved semesters."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "semesterId": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "semesterName": {
                                                        "type": "string",
                                                        "example": "First Semester"
                                                    },
                                                    "semesterType": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "version": {
                                            "type": "string",
                                            "example": "1.0"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "description": "Enter your authentication token",
                "name": "token",
                "in": "header"
            },
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}